#613 refactor: Upgrade the voice-assistant tests
