+++ services/voice-assistant/package.json
@@ -16,11 +16,11 @@
     "scripts": {
         "build:alexa": "jovo build:platform alexa",
         "build": "tsc",
-        "test": "jest",
         "clean": "rm -rf dist build",
         "start:debug": "tsc-watch --onSuccess \"node ./dist/src/app.dev.js --jovo-webhook\" --noClear",
         "start:dev": "node --import=../../register.mjs src/index.ts",
-        "start:prd": "node dist/src/index.js"
+        "start:prd": "node dist/src/index.js",
+        "test": "NODE_OPTIONS=--experimental-vm-modules jest"
     },
     "dependencies": {
         "@jovotech/framework": "^4.5.15",
@@ -38,6 +38,9 @@
         "@jovotech/filebuilder": "^0.0.4",
         "@powerpi/common-test": "^0.0.13",
         "@types/express": "^4.17.11",
+        "@types/jest": "^29.5.3",
+        "jest": "^29.6.2",
+        "ts-jest": "^29.1.1",
         "ts-node": "^10.9.2",
         "tsc-watch": "^6.0.4"
     }
